[#4] [Chore] Refactor
