+++ lib/build.gradle
@@ -39,7 +39,6 @@ dependencies {
     implementation(
             "com.google.android.gms:play-services-mlkit-text-recognition:17.0.0",
             "org.jmrtd:jmrtd:0.7.9",
-            "com.madgag.spongycastle:prov:1.58.0.0",
             "edu.ucar:jj2000:5.2",
             "com.github.mhshams:jnbis:1.1.0",
             "net.sf.scuba:scuba-sc-android:0.0.23"
