+++ lib/src/main/java/co/nimblehq/alpine/lib/model/AdditionalPersonDetails.kt
@@ -1,4 +1,4 @@
-package co.nimblehq.alpine.lib.model.nfc
+package co.nimblehq.alpine.lib.model
 
 data class AdditionalPersonDetails(
     var custodyInformation: String? = null,
