+++ lib/src/main/java/co/nimblehq/alpine/lib/mrz/MrzProcessorResultListener.kt
@@ -1,0 +1,9 @@
+package co.nimblehq.alpine.lib.mrz
+
+import co.nimblehq.alpine.lib.model.MrzInfo
+
+interface MrzProcessorResultListener {
+    fun onSuccess(mrzInfo: MrzInfo)
+
+    fun onError(e: MrzProcessorException)
+}
