+++ lib/src/main/java/co/nimblehq/alpine/lib/nfc/ImageUtil.java
@@ -43,18 +43,6 @@ public class ImageUtil {
         return null;
     }
 
-    public static Bitmap scaleImage(Bitmap bitmap) {
-        Bitmap bitmapImage = null;
-        if (bitmap != null) {
-            double ratio = 400.0 / bitmap.getHeight();
-            int targetHeight = (int) (bitmap.getHeight() * ratio);
-            int targetWidth = (int) (bitmap.getWidth() * ratio);
-            bitmapImage = Bitmap.createScaledBitmap(bitmap, targetWidth, targetHeight, false);
-        }
-
-        return bitmapImage;
-    }
-
     public static Bitmap decodeImage(Context context, String mimeType, InputStream inputStream) throws IOException {
 
         if (mimeType.equalsIgnoreCase("image/jp2") || mimeType.equalsIgnoreCase("image/jpeg2000")) {
